+++ src/bagels/models/split.py
@@ -1,6 +1,10 @@
 from datetime import datetime
-from sqlalchemy import Column, DateTime, Integer, Float, Boolean, ForeignKey
-from sqlalchemy.orm import relationship
+
+from sqlalchemy import Boolean, Column, DateTime, Float, ForeignKey, Integer
+from sqlalchemy.orm import relationship, validates
+
+from bagels.config import CONFIG
+
 from .database.db import Base
 
 
@@ -25,3 +29,9 @@ class Split(Base):
     record = relationship("Record", foreign_keys=[recordId], back_populates="splits")
     person = relationship("Person", foreign_keys=[personId], back_populates="splits")
     account = relationship("Account", foreign_keys=[accountId], back_populates="splits")
+
+    @validates("amount")
+    def validate_amount(self, key, value):
+        if value is not None:
+            return round(value, CONFIG.defaults.round_decimals)
+        return value