+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.3.10
+
+- Configurable hotkey for submit and template
+- Changed default record modal hotkeys
+- Expose additional fields in record creation modal
+- Amount rounding validation
+- Changed default theme to tokyo-night
+
 ## 0.3.9
 
 - Relaxed dependency restrictions
@@ -11,6 +19,7 @@
 ## 0.3.7
 
 - New! New transfer modal now has autocomplete.
+- Include item name in delete confirmation messages
 - Fixed: Templates with similar names will now be selected correctly
 - Fixed: Crash when adding new account